fastapi-app/requirements.txt
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
supabase
jinja2
uvicorn
pydantic-settings

